+++ lib/models/device.js
@@ -26,15 +26,11 @@ import {
 	isId,
 	isNoDeviceForKeyResponse,
 	isNotFoundResponse,
-	getOsUpdateHelper as _getOsUpdateHelper,
-	deviceTypes as deviceTypesUtil,
 	mergePineOptions,
 	treatAsMissingDevice,
 	LOCKED_STATUS_CODE,
-	timeSince,
 } from '../util';
 
-import { hupActionHelper } from '../util/device-actions/os-update/utils';
 import {
 	getDeviceOsSemverWithVariant,
 	normalizeDeviceOsVersion,
@@ -89,6 +85,11 @@ const getDeviceModel = function (deps, opts) {
 	const { addCallbackSupportToModule } = require('../util/callbacks');
 
 	const { buildDependentResource } = require('../util/dependent-resource');
+	const hupActionHelper = once(
+		() => require('../util/device-actions/os-update/utils').hupActionHelper,
+	);
+	const deviceTypesUtils = once(() => require('../util/device-types'));
+	const dateUtils = once(() => require('../util/date'));
 
 	const tagsModel = buildDependentResource(
 		{ pine },
@@ -140,11 +141,14 @@ const getDeviceModel = function (deps, opts) {
 		}),
 	);
 
-	const getOsUpdateHelper = once(() =>
-		getDeviceUrlsBase().then(($deviceUrlsBase) =>
-			_getOsUpdateHelper($deviceUrlsBase, request),
-		),
-	);
+	const getOsUpdateHelper = once(() => {
+		return getDeviceUrlsBase().then(($deviceUrlsBase) => {
+			const {
+				getOsUpdateHelper: _getOsUpdateHelper,
+			} = require('../util/device-actions/os-update');
+			return _getOsUpdateHelper($deviceUrlsBase, request);
+		});
+	});
 
 	// Internal method for uuid/id disambiguation
 	// Note that this throws an exception for missing uuids, but not missing ids
@@ -1021,15 +1025,15 @@ const getDeviceModel = function (deps, opts) {
 					$select: ['id', 'device_type'],
 				}),
 			}).then(function ({ application, device, deviceTypes }) {
-				const osDeviceType = deviceTypesUtil.getBySlug(
+				const osDeviceType = deviceTypesUtils().getBySlug(
 					deviceTypes,
 					device.device_type,
 				);
-				const targetAppDeviceType = deviceTypesUtil.getBySlug(
+				const targetAppDeviceType = deviceTypesUtils().getBySlug(
 					deviceTypes,
 					application.device_type,
 				);
-				const isCompatibleMove = deviceTypesUtil.isDeviceTypeCompatibleWith(
+				const isCompatibleMove = deviceTypesUtils().isDeviceTypeCompatibleWith(
 					osDeviceType,
 					targetAppDeviceType,
 				);
@@ -2517,6 +2521,7 @@ const getDeviceModel = function (deps, opts) {
 		 */
 		lastOnline(device) {
 			const lce = device.last_connectivity_event;
+			const { timeSince } = dateUtils();
 
 			if (!lce) {
 				return 'Connecting...';
@@ -2787,7 +2792,7 @@ const getDeviceModel = function (deps, opts) {
 			// rely on getHUPActionType to throw an error
 
 			// this will throw an error if the action isn't available
-			hupActionHelper.getHUPActionType(
+			hupActionHelper().getHUPActionType(
 				device_type,
 				currentOsVersion,
 				targetOsVersion,
